+++ void_logging/rocket_league/state_metric_providers/__init__.py
@@ -1,3 +1,3 @@
-from ._state_metric_providers import GoalMetricSharedInfoProvider, EpisodeLengthSharedInfoProvider
+from ._state_metric_providers import GoalMetricSharedInfoProvider, EpisodeLengthSharedInfoProvider, GoalScoreSpeedSharedInfoProvider
 
-__all__ = ["GoalMetricSharedInfoProvider", "EpisodeLengthSharedInfoProvider"]
+__all__ = ["GoalMetricSharedInfoProvider", "EpisodeLengthSharedInfoProvider", "GoalScoreSpeedSharedInfoProvider"]