+++ void_logging/rocket_league/state_metric_providers/_state_metric_providers.py
@@ -1,8 +1,10 @@
 from typing import Dict, Any
 
+import numpy as np
 from rlgym.rocket_league.api import GameState
 
 from ..metric_providers import StateMetricSharedInfoProvider
+from rlgym.rocket_league.common_values import UNREAL_UNITS_PER_METER, BALL_MAX_SPEED
 
 class GoalMetricSharedInfoProvider(StateMetricSharedInfoProvider):
     """
@@ -32,7 +34,7 @@ class GoalMetricSharedInfoProvider(StateMetricSharedInfoProvider):
 
 class EpisodeLengthSharedInfoProvider(StateMetricSharedInfoProvider):
     """
-
+    Sets the episode length inside the shared info if episode is finished
     """
 
     @property
@@ -52,3 +54,15 @@ class EpisodeLengthSharedInfoProvider(StateMetricSharedInfoProvider):
         self._episode_length += 1
 
         return None # IDK why python doesn't just...do that on its own ?
+
+class GoalScoreSpeedSharedInfoProvider(StateMetricSharedInfoProvider):
+    @property
+    def metric_name(self) -> str:
+        return "Misc/Goal speed (kph)"
+
+    def get_metric_value(self, state: GameState, shared_info: Dict[str, Any]) -> float | None:
+        if state.goal_scored:
+            _mps_speed = np.linalg.norm(state.ball.linear_velocity) / UNREAL_UNITS_PER_METER # uu/s to m/s
+            _kph_speed = _mps_speed * 3.6 # m/s to km/h
+            return _kph_speed
+        return None